Adding delete property to ItemInfo.
This fixes a bunch of problems where the frontend checks the ItemInfo
object to see if it has a device property and does stuff based on that.
Since all non-device items had no property, all those checks would
bomb out.

I think this is a stopgap fix so that Miro isn't handicapped until
Paul redoes some of the DeviceItem stuff.