Implemented test skipping; fixed tests for Windows.
There are now two test skipping decorators in framework.py.
only_on_platforms allows you to specify that a test only
run on specified platforms.  skip_for_platforms is the inverse
and allows you to specify which platforms should skip the test.
Tests that are skipped are specified at the end of the run
so that it's clear what got skipped and why.

I removed the code I hacked for test skippage a while back because
it sucked and it's redundant now.

Additionally, I fixed some tests that were failing on Windows.