Fix #15749.
Add a new way to configure a version of Miro tailored for the Mac App Store.
Use it so as to change the software update to direct users to the Mac App
Store to get their updates if it is a Mac Store edition.

Additionally, disable auto update on Mac OS X if running on App Store edition.