Stop playback before starting it again (portable part of #15752)
This was a real bug because the previous code caused would totally mess up the
signal handling for that case.  I haven't tested it, but my reading of the
code tells me it would never disconnect the old handler, and connect then
disconnect the new handler.  I may be wrong about that, but I'm sure it was
broken in some way.