(windows) Overhaul of logging, debugmode, testing, ...
Created a mirodebug target which is a py2exe console application that's otherwise
identical to the Miro target.
Added a --debug flag that causes Miro to log to stdout as well as the rotating log.
Removed the bdist_test command and folded the functionality into the bdist_miro
command.
Then by accident fixed a lot of end-of-line whitespace issues--I have Emacs set
to remove them automatically.