Made feedparsertest more lenient and easier to debug.
Since EOL sequences don't matter in parsed feed output, we remove possible
differences between output and parsed output on Windows.
Also, we manually compare the strings so that if there are differences we
show the context and not the whole string--there's a lot of feedparser
output and this will make debugging easier.