Bug #15773.  Fixes iconcache problems on Windows.
next_free_filename was changed in November and now opens the file
and returns a file object for it.  One of the problems this creates
is that there's now a file in the location that we want to rename
the temp file to.  On Windows, this kicks up an OSError which is
silently absorbed and thus nothing comes out of iconcache.

Fixing that brings up the second problem.  For some reason, when
saving the info["body"] to the temp file, the data was being treated
like ascii data.  On Windows, ascii data has newlines converted.
Thus all the images were corrupted.  I don't know understand why
next_free_filename is returning a file object for a file opened
as an ascii file.  It's certainly not apparent in the code.  Closing
that file and reopening it as wb fixes the corruption problem.