Improved the OS X garbled UI a bit more.
Made it so we process size requests right after we process an event.  This
makes updates happen quicker and improves things generally.

The main thing this fixes is the tablist when you startup the application.
It's still not perfect, but now it's more that not all the display updates
rather than things are writing over each other.

Added an event-processed signal to OSXApplication