Blind fix for #16070
Changes to InfoList make us need to catch a KeyError rather than ValueError
there.

Changed our error recovery so that we still try to open the link.  #16070
makes it clear that not finding the feed is a semi-common case since it
happens when an item is moved to the manual feed.