Partial fix handling videos from command line/contetx menu
Added back PlaybackManager.start_with_items().  There's still a use for it
when we want to play a non-changing item list (examples movies from the
command line, or play just this item).

Refactored how we handle ManualItemsTracker to avoid doing a round-trip just
to start a movie where we already have the ItemInfo for.