Fix for #16113
The issue is that the GTK code relies on removing nodes from our info list in
reverse order.  To do this we sort them, however if the sort keys were equal,
then what multiple orderings were possible.  If we sorted them one way when
inserting, then sorted them a different when removing, things blew up.

To fix this, add a tiebreak so that nodes always compare greater or less than
the other.