Implemented #15646
Changed right arrow to move to the next track instead of just skipping forward
in time on the current one.  Ditto for left arrow.

Allow playback keyboard controls to work with detached playback/audio items.

Changed GTK so that the key-press event for windows is hooked up to an
eventbox rather than the window itself.  This makes it so that we get
key-press events after other widgets have had the chance to handle them,
rather than the other way around.  In particular, this fixes hitting the
right/left button in text entries.

Fixed Cocoa Event handling to work with the new keyboard controls.  Instead of
giving our wrappers the chance to handle an event, then passing it on to the
cocoa views, do it in the opposite way.  This way if the user has a text entry
focused, that will handle the event.