Bug #16116.  Reworks fastresume data handling.
Miro no longer sends fastresume data across the wire between the downloader
and the main process and no longer stores fastresume data in the database.

The downloader now saves/loads fastresume data based on the info_hash
of the torrent handle.  Fastresume data is stored in the support dir
in the fastresume/ directory.