Better key-press handling on OS X.
Before we would give all cocoa responders a chance to handle the event, then
giving it to the wrappers.  Now we follow a chain like this:
  wrapper for firstResponder
  firstResponder
  wrapper for firstResponder.nextResponder
  firstResponder.nextResponder
  ...

This is what happens on GTK and seems like the nicest way of handling things.
wrappers get to handle events before the NSViews they wrap, but not if a
NSView that's before it in the responder chain handles it.