Don't authenticate SSL Certs #15180
Many feeds in the wild will fail the test and it's not like we are doing
anything super secret here.  If an attacker manages to fake a cert the best
they can do is trick you into downloading some videos.  Plus if we wanted
great security, the list of CA authorities that libcurl trusts may not be the
one we want to use.