Tweaks to ItemListController's handling of changes.
Added methods to call model_changed() and start_bulk_change() for all item
views.

Call model_changed() for all views when updating the sort.  The sort affects
views that aren't display, so I think this is necessary.

Call start_bulk_change() when we are adding/removing lots of items.  This
usually is changing filters or sorting.  In any case, the reason to avoid it
is to keep the scroll position, but when many things change, it's not useful.