Bug #15675.  Fixes a lot of bad exception handling.
In some places, I was able to lock down the exception handling
appropriately.  In other places, I had to settle for making it a bit
better because there was so much code in the try block and most of the
function calls are poorly documented and it's a big project to lock
it down further.

Regardless, this should be better than what was previously there.