Get rid of broken InfoList code for no sorting.
I added that code for playlists, but playlists actually do use a sort.

The code itself is busted because we rely on being able to remove ItemInfos in
reverse order on GTK.  Without a sort we weren't doing this, which lead to
ticket #16248.

Some tweaks for sorting playlists, it's still not perfect though.  See #16278
for the discussion.