Made list/standard view share sort order (#16278)
This was actually sort-of happening, because they shared an ItemList.  This
change makes it so the code is no-longer fighting each other.

Added code in to not show any sort in standard view if we are sorting on one
of the list view columns that doesn't have an equivalent in standard view.

Simplified sorter code now that we no longer need to worry about
multiview_sorter being a special case.