ngettext speedup
We were spending somewher between 5% and 10% of our time in list view
rendering making calls to ngettext.  Now it doesn't register in my profile
code.

Create a Translation object in gtcache.init() and use that.  A huge amount of
time in ngettext was spent finding the .mo files.

Using the new API allows us to be much more flexible in initializing gettext.
This allows us to change languages in unittests without setting env variables.

Removed code that calculated codeset since we don't do anything with it.