Bug #16339.  Fixes write_resume_data kicking up RuntimeError
This is not an optimal fix, but it's probably good enough for now.  The
better fix would be to rewrite the fast resume data code to use
save_resume_data which uses libtorrent alerts.  However, I think that'll
take a chunk of time and it's better to push that off to the next
release.