Bug #16321.  Fixes textbox saving in pref panel.
Somehow I screwed up and implemented the wrong method.  This implements
a better one, but it still doesn't handle the case where the user
changes the value and either closes the pref panel or shuts down
Miro.  In either case, textDidEndEditing doesn't kick off.  That still
needs to be fixed.