Fix for tableview hover handling on OS X.
The old code was really slow and I believe it caused #16152.  Now, insted of
trying to create a tracking rect for every cell, we create one for the entire
view and calculate what cell we're in in the mouse moved handler.

Changed code to use TrackingAreas now that 10.4 isn't supported.

Changed our NSWindow class to not accept mouse moved events.