One more fix for #16225
When the tablist selection changes, don't clear the whole display stack.
Instead, just replace the bottom display.  This way if video is playing on top
of the display for the tab, it will continue.

Added some code to deal with removing displays that aren't on the top of the
stack.