Implemented most of #15949
Added video count to the top-right corner of standard view.

Ton of changes to ItemRenderer.
  - Added lots of new images.
  - Gave all images the prefix "item-renderer".
  - Use a dict for storing images since there are so many of them.
  - Split out code to draw the emblem.

Some fixes to draws_selection.  The previous workings were totally mysterious
to me.  There was the draws_selection attribute and set_draws_selection, and
they seemed to do two related, but different things.  You had to set one to
True and one to False to do what you wanted, which is totally nuts.

Shortcut all code for the converting renderer, I'll re-implemented that later.