Make entire standard view dark, even if there are not a lot of items.
Made it so we always use the same background for standard view, and it's set
to the standard view background color.

Finally removed some old OS X scroller code.  I'm sure it was a good idea at
the time, but I don't get it now.  Plus, removing it fixes the background.