Fix for list view scroll jumping.
In list view whenever we updated the model, the scroll position would jump
around.  I'm not sure exactly what the bug was, but it was due to our Scroller
seeing that it's children changed, then calling invalidate_size_request().
That combined with the TableView calling invalidate_size_request() caused
issues.  There's no reason at all for the scroller to cal
invalidate_size_request() because it's size is not dependent on its children.

Note: there is still a general TableView bug that causes jumps in the scroll
position.  I'm pretty sure the problem code is in set_scroll_position().
If we comment that out, then things seem to work.  I'll file that as a
separate ticket.