Fix for item license info in the details pane.
Before we were calling widgetutil.align_left() every time we added a widget.
However, that function creates another widget and sets the paramater as the
child.  That broke for many reasons.

This replaces geoff's fix.  I think it works better on GTK because it avoids
giving a child widget multiple parents.