Split up ItemRenderer
Now we have several classes:
  - ItemRendererBase is the base class for ItemRenderer
  - ItemRenderer determines what should be visible for an ItemInfo
  - ItemRendererCanvas positions/draws the elements ItemRenderer adds

The idea is to make subclassing easier.  We can subclass ItemRendererBase and
change the logic, but still use the same layout/drawing code.  The big example
of this is conversion item renderers.

Moved the code to create/cache ImageSurfaces out of ItemRenderer and into a
shared module function.  This makes it simpler for all classes to use the
images.