16654.  Implements Import Media dialog.
It's based in large part on the First Time Dialog, but I ended up
duplicating a lot of the code.  While that's kind of lame, it's better
because there are some things in the First Time Dialog that we don't
want to do.

Anyhow, it's probably a good first pass.