Make sure that strftime_to_unicode() returns unicode, in most cases.
Strftime_to_unicode() kept returning things as str because the gtcache.codeset
was not initialized, which broke for various reasons when you tried to combine
unicode strings with it and leads to unicode exceptions from Python when it
tried to use the ascii code to munge to unicode.

Still not correct because in the error cases we still return 'str', that's
not correct.