Bug #16573: sources picking up old icons from previous sources
Our ImagePools were caching an icon for 'favicon.ico', which is what all the
guide icons are called.  Even though we updated the icon on disk, the cache
didn't pick that up.  For now, we just don't use the cache for the Guide icons,
but it's probably better to fix this by clearing the cache in the future.