Do our URL checking earlier on linux and OSX.
Doing it this way allows us to catch links before the browser tries to load
them.  This fixes magnet links.  Before, our browser would see a magnet:
scheme and bail before we could intercept the link.

Moved code that actually needs to use the mimetype to a new method.

On windows we already were handling links like that, so no need for changes.
It also means that should_load_mimetype() is never called on that platform.