Improved new search feed dialog.
This makes it so we don't need to put it in the tab cycle for keyboard focus,
since the activating the menu item + hitting enter does the same thing as the
button.

Split up the new search feed code.  Put it al inside a class to make it easier
to track the widgets we create.

Make the save as podcast button unfocusable.  Now that the dialog does the
same thing, we don't need to have it in the tab cycle.