Made code less weird in addfilesdialog.py
The issue was that the choice page was overflowing the height.  Using
align_bottom() hacked around some of the issues, but it's cleaner to use pad()
and give enough height for all pages.

Making the height 230px, which should give enough room to grow with font size
changes.

Log when content widgets exceed their available size on OS X.