Keyboard focus fixes for OSX (#16932)
Immediately re-layout containers when their children is changed.  This makes
sure that we can focus widgets once we add them to the widget tree.

Don't re-set the same document view for our scrollers.  That causes the view
to be temporarily removed from the window and it loses first responder status.

Don't try to set a view to be the first responder if it's not inside a window
yet.