Allows shutting down Miro with ^C.
KeyboardInterrupt would propagate out of the gtk.main() loop and
encounter nothing and cause Miro to hang because the main loop
was over, but the rest of it was hanging around.  This fixes that by
handling KeyboardInterrupt and calling do_quit if it's encountered.
That creates the minor problem that do_quit eventually can call
gtk.main_quit() which gets fussy if it's called outside of the main
loop.  So I fixed that, too.