Simplify table selection code
Removed the code to save and restore selection auto-magically.  Moved
the code from the tableselection, to the classes that use the tables.
The new system is less general, but hopefully more reliable.  For
example we now try to restore selection once, rather than on every
model change until it works.

For item lists:  We save the selected ids when we leave the view, then
restore them when we return and get the initial item lists.  We also
save/restore the selection when we make changes to the model.

For tab lists:  We were basically using this scheme before, so not much
changes were needed.

- Refactored ItemListController, so handle_item_list() always gets
  called, even if our item tracker is initially populated.
- Generally simplified the tableselection API.  Now that we don't deal
  with restoring selection here, there's a lot less exceptions to
  worry about and we don't need the strict and force parameters
- Removed select_path method, we were only using on GTK, we can just
  define it there.
- Store selection as a list of strings or ints.  Added schema type to
  handle data that can be different basic types.
- Removed a bunch of _save_selection() calls in tablistmanager.py.  I
  couldn't tell exactly what they were doing, but it seemed not needed
  anymore.  Plus Kaz tells me that they aren't needed after his recent
  round of changes.