16873.  Swaps out table for links.
This removes the label - Go table and implements hyperlink things
as Nick suggested earlier in the bug.
I did this because I couldn't get the text of the label to align
correctly with the Go button without fiddling with it longer and
the Go button wasn't working for me visually anyhow.