Fixes for crash reporting (#17037)
Changed the way we handled the call stack.  If we have an exception, the call
stack is pretty useless, since the exception's traceback is much better.  If
we don't have an exception, use the call stack to fake one so that bogon
deflector parses the report correctly and we don't get empty reports.