Hacky fix for test.sh not taking multiple args
I didn't feel like chasing this down, but test.sh doesn't pass all the
command line arguments to run.sh.  So doing something like:

    ./test.sh -v moviedatatest

passes the -v to run.sh and totally loses moviedatatest.

This fixes that in a hacky way by explicitly passing arguments, but it
only passes the first three because that's as many as I pass.

It should get fixed for real at some point.