16781.  Fixes gst_extractor thumbnailer from hanging.
For some videos (probably not many--I haven't seen this until today),
the seek happens, but the buffer_probe never kicks off.  I didn't spend
a ton of time figuring out why this happens, but when it does, it
causes the gst_extractor to just hang out and never complete because
the completion code is in the buffer_probe_handler_real.

Thus I added a 2 second timeout to ensure that gst_extractor completes
whether or not the buffer_probe kicks off.