fix 17054: add files dialog broken on OS X
addfilesdialog is depending on the clicked signal being sent with an
event parameter; this event parameter is a GTK-specific object. I'm
removing the event parameter from CustomControlBase.

This class has a lot of descendants so it's hard to be sure that there
isn't anything depending on the old behavior, but anything that did
would likely have been broken on OS X.