fix 17055: prevent tablist double-toggling
expander-size does not take into account horizontal-separator or
EXPANDER_EXTRA_PADDING, leading to a gap of several pixels between the
area tableview thinks is part of the expander and its actual right edge,
causing double-toggles when the expander is clicked and tableview also
forwards the signal. Using get_left_offset gives the correct position of
the expander's edge.