Fix for #17059.
Part of the issue was that for certain files VLC was outputting a ton of
warnings.  The program never terminated because it would fill up it's output
buffers and wait to send more.

Use a temporary file to catch STDOUT and STDERR from the movie data program to
avoid this.