Bug #17063: Browser titlebar glitch when window is resized
OS X doesn't seem to remove the old child if we call set_child() during size-allocated.  To work around this, we explicitly remove the old child before re-showing the new child.  We could try to be more clever, and only do the remove/hide if the widget has changed, but this seems okay for now.