17120. Fixes unicodedecodeissue
On OSX and Linux, it looks like params["save_path"] is a string, but
on Windows it's a unicode.  It's probably a PlatformFilenameType.
The problem occurs because we're converting it to a utf-8 encoded
string.  We should only be doing that if it's a unicode (i.e. on
Windows).

I threw a big FIXME to look into the problem more closely post-4.0.