Fixes for item list empty mode (#16970)
For searches, we only want to show the empty list mode before a search starts,
or if the user clears a search.

For other ItemListControllers, removed redundant code to calculate empty list
mode.  Calling set_search() will result in the items-changed signal being
emmitted, and we check for empty lists there.