Bug #16728: devices not detected on Windows
It seems that some version of Windows return different data for the various
device ID function calls.  It was working for me on Windows XP, but not on
Windows 7.  This checkin makes it work for me on both platforms.  I've also
added some very verbose debugging support (currently disabled) so that if there
are future problems I can get more useful data.

As a side-effect, Windows no longer uses the USB vendor/product IDs; we get the
USB name out of the registry.  In a future checkin, I'll remove support for
that method of looking up devices for simplicity.