VLC error handling fixes (#17189)
Don't call our errback while we are inside setup_item().  We call setup_item()
in response to the errback, so we can get a nested signal error if we have 2
items in a row with errors.

Reset our state before calling errbacks/callbacks.  If the handler calls
setup_item(), then we don't want to overwrite the attributes.