partial fix 17153.1: save/restore scroll on view change
There is a complex problem on OS X in the interaction between Scroller
and tableview that causes Scroller to forget its position when it is
hidden.

This works around that by restoring the scroll position on every view
change. It's certainly not ideal, but this should fix the problem
without the risk of side effects involved in changing
Scroller/tableview signals, and on GTK where it's already working this
has no effect.