fix 17153.1 for Displays
OS X loses Scroller positions whenever they're hidden. This is a
workaround for changing Displays that is similar to the earlier
workaround for Views: after switching to a display, once scrolling is
possible, we set the scroll position. I think it is possible to avoid
losing track of position in the first place (as we do on GTK), but
would require substantial changes - so this is a temporary solution.