Moves the timeout in the gst_extractor
Janet has a .flv video where gst_extractor hangs on the thumbnail side
of things.  It looks like the bits of the thumbnail pipeline all change
state to PAUSED, but the pipeline itself never does.  I'm not really sure
what to do with that--we really want to take the thumbnail only if
the pipeline is ok.

I had a 2-second timeout in an earlier change.  I moved it for this change
and increased it to 3 seconds.  If it takes longer than 3 seconds for the
thumbnail pipeline to get to PAUSED, we cut our losses.