17176. Fixes problem where torrent is None
This is sort of a blind fix since I don't have any steps to reproduce
the issue.  But what I did was add a check for self.torrent is None
and in that case, just return since there isn't any fast resume stuff
to save.

Also, I consolidated the "if this, then return" stuff to the top of the
method.