Bug #16963: do a better job cleaning up after processes
When we killed processes before, we weren't actually cleaning up after them all
the way.  This changeset uses the support in the subprocess module to kill and
cleanup (via wait()) the subprocesses.

Also, this nearly removes the necessity for a platform-specific
kill_process(pid) function, but it's still used by the download daemon.  At
some point, we should refactor that code in order to use more Pythonic
subprocesses.