Fix for #17020
Keep the keyboard cursor in the same location when we switch from list view to
standard view.  This will ensure that we don't get in a situation where the
row with the cursor is not selected.

Only implemented it on GTK, but OS X seems to do things alright as-is.