Better handling of low width for titlebars (#16554 and #17269)
Renamed "small monitor mode" to "small width mode" and have it work
automatically instead of using a pref.

Took the idea from z3p's progress bar and extended it to titlebars.  We
calculate the width we need for small width mode and set our size request to
that.  We calculate the width needed for normal mode and if we are allocated
less size than we go into small width mode.

Made the resume button hide in small width mode, rather than using the old
system of giving it a ton of space, but only draw it on the left side.  This
fixes #17269.