fix 17192 - play_pause broken
play_pause() is broken after stop(), and it seems it has been for a
while. It's possible that it is behaving as intended, but should be
called pause_unpause and should be more assert()ive about it. This does
not fix play_pause - this just avoids using it in the situations where
it doesn't work.

Bug #17339 refers to removing this workaround and fixing the underlying
problem.