Don't let feedparser hog so much CPU (#17216)
Added some logic to how we call feedparser.  Instead of just sending
feedparser.parse calls to the worker threads, we queue them so that we only
send one at a time and sleep a little between each call.

This makes the UI freeze less when we are doing a bunch of feedparser updates.