17295. Comments out "feed can't update--remove it?" dialog
It's likely Miro will be used more often when someone isn't connected
to the network.  We don't have time to implement an offline mode, so
we're commenting out the "feed can't update -- remove it?" dialog which
pops up for every feed when it updates when the computer is not connected.

When we implement an offline mode, we can re-add this dialog.