Move check_media_file() back to setup_new() where it belongs
I moved it out in 9df6424 to try to fix #17344, but we depend on it being
there in a lot of places.

I think this solution is actually nice and clean.  It's nice to always run the
same cleanup code for any error.

I think it fixes all the issues, which were: make sure that we leave
incomplete_mdp_view, remove FileItems for deleted files, and call soft_fail.
Also, I realized that we should probably call path_processed to keep the
metadata progress count right.