Fix for #17272
Move the handling of config changes to the ItemListController.  This way we
can disconnect the signals when we're done with the controller.

Made us track config changes as long as the controller is around, rather than
just as long as it's displayed.  That makes more sense in general to me.