make sure that device JSON keys are always Unicode
We weren't being careful about making sure that we accessed the JSON dictionary
with Unicode keys, which can lead to problems as we convert between Unicode and
strings.  This code adds two checks to __getitem__ and __setitem__ which verify
that keys are always Unicode, and then fixes all the issues that appeared with
those checks.

I believe this will resolve #17310, #17337, and #17359.