preempt gtk iter assertion with our own assertion
Related to #17362 - when a GTK assertion that an iter is valid fails,
Miro bails without a dialog or any kind of traceback.

This changes GTK TableView to wrap any use of get_path in a method that
also asserts iter_is_valid. Also, adding a noop (for now, anyway)
InfoListModel implementation of iter_is_valid.

The result of this change should be that the hard crash seen in #17362
becomes an assertion failure.