17391. Reworks search page in first time dialog
Overhauls search page in first time dialog so that it immediately
starts searching when the user clicks "Next >" to the page and also
has a cleaner ui for users to understand the different searching
states (searching, finished searching, cancelled search).

This also cleans up some odd behavior when going back and forth between
pages.  Now if you click on the No radio button in the Find Files page,
it clears the gather_media_files list.

If the user cancels the search, then it now clears the
gathered_media_files list.