Fix for torrent folder migration (#17393)
The issue was that we were calling next_free_filename() and that creates a
file to use for us.  Then calling shutil.move() with a directory is a source
and a file as a destination failed.  Changed to use next_free_directory() to
make a directory that shutil.move() can work with.