16436#c26: run MDP for all OGGs
Mutagen calls some OGG videos audio; we need to run MDP for anything
with a .og* extension (except for .oga), to avoid miscategorizing a lot
of ogg videos. OGG files also should not be playable until MDP has run
for them, because otherwise right after downloading a video, it would
play as audio.