Potential fix for #17429
I couldn't get the unittest failure on my system, it's probably an SQLITE
version thing.  However, I believe this code will fix it.

The unittest failure also pointed out a real problem.  If we can't run the
PRAGMA statement, what do we do?  I made it work the same as if we couldn't
get the last_id from the database, throw the database away and restart.