17205, 17427. Removed a failing unit test.
I looked at the test_media_with_mdp tests and I'm not convinced that we're
testing Miro functionality so much as we're testing gstreamer functionality.
That's great and all, but it's pretty outside our powers to fix any issues
that come up like the ones in these two bugs.

For now I'm ditching the test that's failing.