17362: don't leave invalid iters in iter_map
handle_jettison_tabs deletes some of the tablist's iters until
setup_list re-creates them; in the time between jettison_tabs and the
new setup_list, the iter_map was full of crashy iters.

Adding a note to TabList describing some rules for avoiding iter-related
crashes and segfaults. After this change, I think we follow (0) - and I
have verified that we are following (4) - so I believe we are now clear
of segfaults. I'm not totally sure we're following (2) and (3), which
could lead to KeyErrors and minor misbehavior, respectively - but there
haven't been any recent reports of either.