Handle weird windows directory states while moving downloads (#17456)
The problem is that because we're using GFileMonitor to monitor the movies
directory and it's subdirectories, we keep directory handles to all those
paths.  When we delete the last item in a feed, we try to delete it's folder,
and that causes the folder to go into a half-deleted state.  It makes it so
os.path.exists() returns False, but os.path.makedirs() can't re-create it.

The real fix for this is #17460, but I think this will do for 4.0