17442: need to use MDP to identify DRM
Mutagen can only identify whether a file "might" have DRM. The strategy
I'm using here is to let mutagen set has_drm if the file could possibly
be locked, and tell MDP never to skip over files with has_drm set. Then
if MDP can get a duration for the file, it unsets has_drm.

In between when mutagen look at the file and when MDP looks at the file,
we are in a state where all we know is that the file "might" have DRM -
because mutagen sets has_drm, is_playable returns False (we don't want
to try to play anything that might have DRM - #16405) - but when making
an ItemInfo we set has_drm=False for the ItemInfo until MDP has looked
at it. This way we don't try to play anything we can't, but also don't
confusingly temporarily mark files from iTunes as DRM'd while we're
importing.