17461: don't disconnect shuffle/repeat while playing video
Previously an ItemListController has been connected to the
shuffle/repeat/stop signals when it is the current display. This is not
quite right: when a VideoDisplay is on top, the itemlistcontroller
should still be connected to the signals. I have implemented a 'primary'
display that is the most recent itemlistcontroller to be displayed; the
'primary' display is the same as on_display in most cases, but does the
right thing while playing a video.

I've also added soft_fails to where the repeat/shuffle/stop handlers in
item_list_controller catch the case where a signal we shouldn't be
connected to arrives; if this case ever does occur, I think
shuffle/repeat are not going to be right - I don't think that it's
possible, but if it is soft_failing allows us to identify the cause of
the misbehavior.

This will be a little strange when playing a video that did not come
from the itemlistcontroller that was open before playing the video -
i.e. when using File->Open: in that case, the shuffle/repeat are still
connected to the itemlistcontroller that was displayed before playing
the external video. I think this edge case is less of a problem than
the previous problem while playing any internal videos.