17441: performance marking items watched/unwatched
Now sending one message with an iterable of items to mark as
watched/unwatched rather than send a BackendMessage for each item.

This only improves the time to completion by roughly 15%, but the work
it saves is largely in the frontend, and it does seem to improve
responsiveness during the operation.