Fixes for item creation in item info cache's failsafe load
Before we would just ignore them, but this isn't right because the item will
not be in the info cache.  Now we put it in the info cache, but skip the code
to save the changes, since we save everything after the failsafe load.

Made a similar change for item removals.