Fixed torrent folder paths (fixes #17472)
Made it so that for torrent folder items, self.filename points to the
folder that libtorrent downloaded, instead of the parent directory of the
folder.

This matches the 3.5 behavior, which fixes #17472.

Made next_free_directory() not create the directory that it finds.  This fixes
a lot of weird migration behavior that we used to need hacks for.  I don't
think there will be many problems with this, and besides it's the 3.5
behavior.