Fix for bulk delete of children and parents together (#17481)
The crash was happening because we were using BulkSQLManager.  That causes the
items to not be removed from the database until we finish the bulk
transaction.  So we would delete the parent, and that would delete the
children.  Then we would go to delete the child and it would still be in the
DB.  This caused us to try to delete the child twice and everything broke.

The fix is to keep track of which items are going to be deleted, and check
that in our bulk_delete() method.  I made it work like the existing code for
inserts, except I changed the sets to track IDs which simplifies the work.