17370. Amazon downloads fail to start on OS X/Linux
Amazon is looking for 'Safari' in the user agent string.  Since we stripped it
out, they weren't serving us the Javascript we needed to do much of anything.

NB: On OS X, I couldn't figure out how to get the 'Safari/531.2+' number for
real, so I hard-coded it.  view.applicationNameForUserAgent() returns None.