Better fix for #17483
We had the _is_playing_filtered_item attribute to handle cases like that, but
the issue was that it would only pay attention to what was happening at the
ItemTrack level.  This meant it could handle items being removed, and items
being filtered by searches, but not items being removed by the filter buttons.