Ultimate fix for #17483
This fixes the stuff before and also the crash in comment 18.

The issue was that we don't get a signal when we change filters for ItemList
so we had no way of updating _is_playing_filtered_item.

I made _is_playing_filtered_item() a method, so it should always be correct
and up-to-date.