17515. Fixes ui problems when changing language in first time dialog
There were two different problems.

The first is that we have a bunch of code that translates strings at
import time rather than when the objects are instantiated.  I tried
to fix this with the least amount of code changes as possible, so it
looks stupid and it's redundant, but it works.

The second is that we were using the translated string in comparisons
to figure out which display to show.  But the lhs string was translated
at import time and the rhs string was translated at the time the
comparison was being made and thus they'd be different.  I tried
to fix this by re-using the type on the statictab classes.

There are several things that I ran into that make the code suck:

1. translating at import time
2. conflating class and instance attributes and using them
   interchangeably willy-nilly
3. overuse of the word "type" which is also a built-in
4. if blocks that compare translated strings

A lot of that is stuff to fix another day.