fix 17513 (crash clicking Connect during startup)
There is apparently a very short window of time between when the
ConnectTabRenderer is created and when pack_bubbles runs, during which
clicking on the tab causes an exception because hbox has not been set
at all.

I haven't been able to reproduce this consistently, so this fix isn't
fully tested - but it definitely fixes the AttributeError, and shouldn't
affect anything else.