17329: certain glyphs break ItemRenderer title
The problem is that some glyphs like 0xFFFD (placeholder for lost
character, usually because of an encoding mishap; looks like a black
diamond) can be slightly taller than the way we currently determine
line_height: ascent + descent is not guaranteed to be as tall as the
tallest glyph in the font, and 0xFFFD exposed this problem because it's
particularly tall.

The solution for now is to add 1px when determining line_height on GTK.
Filing a 4.1 bug for finding a more accurate way of getting actual
height: 17539; I don't see any straightforward way to do so.