fix 17557: handle any enc for cover art mime type
Apparently some files have surprising characters / encodings in the
mime-type field for their cover art. Implemented _text_to_ascii, which
takes a str or unicode containing arbitrary data and returns a str with
only the ascii characters in the input - for mime types, this seems the
most likely way to get a meaningful result when this happens.

I have not tested this on an actual file because I don't have a test
case and only a broken metadata editor would be capable of creating one,
but the error is easy to simulate by running the line where it shows up
("{0}".format(text)) with non-ascii unicode data. I have tested
_text_to_ascii on a variety of types of string data, and it handles
every type of case I can think of.